LayoutTests/http/tests/security/dataURL/xss-DENIED-to-data-url-sub-frame-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL data:text/html,<html><script>onload = function() { parent.postMessage('LOADED', '*'); } </script><body><p id='accessMe'></p><p>Inner iframe.</p></body></html> from frame with URL http://127.0.0.1:8000/security/dataURL/xss-DENIED-to-data-url-sub-frame.html. Domains, protocols and ports must match.

This tests that the main frame can't access the contents of an iframe that contains a data: URL loaded page


PASS: Cross frame access to a data: URL was denied.


--------
Frame: 'aFrame'
--------
Inner iframe.
